Remove unused update_sds_version.py
The script update_sds_version.py isn't used anywhere anymore.
The function move_patches_up_to_date_to_source_data_stream_component
that is still used is moved to the script that uses it (compose_ds.py).
